Initial GIT commit...
Project is claimed to be GPL 2.0 on Sourceforge.  Going to clean it up
and re-provide it under GitHub with a CMake build system since the
other's quite broken on several fronts, relying on things like Google
GFlags' being up on Google Code, etc.

It now compiles instead of cowpiles.  We'll see how well this works
against the Expression I have in hand.  Should be "good" even if
it desperately needs some cleaning up on docs, build, etc.